+++ OUDS/Core/Components/Sources/Controls/Checkbox/OUDSCheckboxItemIndeterminate.swift
@@ -74,15 +74,15 @@ import SwiftUI
 ///     // The default layout will be used here.
 ///     OUDSCheckboxItemIndeterminate(selection: $selection, label: "Hello world", isReadOnly: true)
 ///
-///     // A leading checkbox with a label, and an helper text.
+///     // A leading checkbox with a label and a description as helper text.
 ///     // The default layout will be used here.
-///     OUDSCheckboxItemIndeterminate(selection: $selection, label: "Bazinga!", helper: "Doll-Dagga Buzz-Buzz Ziggety-Zag")
+///     OUDSCheckboxItemIndeterminate(selection: $selection, label: "Bazinga!", description: "Doll-Dagga Buzz-Buzz Ziggety-Zag")
 ///
 ///     // A trailing checkbox with a label, an helper text and an icon.
 ///     // The reversed layout will be used here.
 ///     OUDSCheckboxItemIndeterminate(selection: $selection,
 ///                                   label: "We live in a fabled world",
-///                                   helper: "Of dreaming boys and wide-eyed girls",
+///                                   description: "Of dreaming boys and wide-eyed girls",
 ///                                   isReversed: true,
 ///                                   icon: Image(decorative: "ic_heart"))
 ///
@@ -100,7 +100,7 @@ import SwiftUI
 ///
 ///     // A leading checkbox with a label and and icon but with the icon flipped vertically
 ///     OUDSCheckboxItemIndeterminate(isOn: $selection,
-///                                   labelText: "Cocorico !",
+///                                   label: "Cocorico !",
 ///                                   icon: Image(systemName: "figure.handball"),
 ///                                   flipIcon: true)
 ///
@@ -115,7 +115,7 @@ import SwiftUI
 ///     @Environment(\.layoutDirection) var layoutDirection
 ///
 ///     OUDSCheckboxItemIndeterminate(isOn: $selection,
-///                                   labelText: "Cocorico !",
+///                                   label: "Cocorico !",
 ///                                   icon: Image(systemName: "figure.handball"),
 ///                                   flipIcon: layoutDirection == .rightToLeft,
 ///                                   isInversed: layoutDirection == .rightToLeft)
@@ -148,7 +148,7 @@ import SwiftUI
 ///
 /// ![A checkbox item component in light and dark mode with Wireframe theme](component_checkboxItem_Wireframe)
 ///
-/// - Version: 2.3.0 (Figma component design version)
+/// - Version: 2.4.0 (Figma component design version)
 /// - Since: 0.12.0
 @available(iOS 15, macOS 15, visionOS 1, watchOS 11, tvOS 16, *)
 public struct OUDSCheckboxItemIndeterminate: View {
@@ -169,8 +169,8 @@ public struct OUDSCheckboxItemIndeterminate: View {
     ///
     /// - Parameters:
     ///   - selection: A binding to a property that determines wether the indicator is ticked, unticker or preticked.
-    ///   - label: The main label text of the checkbox.
-    ///   - helper: An additonal helper text, should not be empty
+    ///   - label: The main label text of the checkbox, must not be empty
+    ///   - description: A description, an additonal helper text, should not be empty
     ///   - icon: An optional icon
     ///   - flipIcon: Default set to `false`, set to true to reverse the image (i.e. flip vertically)
     ///   - isReversed: `true` of the checkbox indicator must be in trailing position,` false` otherwise. Default to `false`
@@ -181,11 +181,11 @@ public struct OUDSCheckboxItemIndeterminate: View {
     ///   - hasDivider: If `true` a divider is added at the bottom of the view, by default set to `false`
     ///   - action: An additional action to trigger when the checkbox has been pressed, default set to `nil`
     ///
-    /// **Remark: If `label` and `helper` strings are wording keys from strings catalog stored in `Bundle.main`, they are automatically localized. Else, prefer to
+    /// **Remark: If `label` and `description` strings are wording keys from strings catalog stored in `Bundle.main`, they are automatically localized. Else, prefer to
     /// provide the localized string if key is stored in another bundle.**
     public init(selection: Binding<OUDSCheckboxIndicatorState>,
                 label: String,
-                helper: String? = nil,
+                description: String? = nil,
                 icon: Image? = nil,
                 flipIcon: Bool = false,
                 isReversed: Bool = false,
@@ -199,8 +199,12 @@ public struct OUDSCheckboxItemIndeterminate: View {
             OL.fatal("It is forbidden by design to have an OUDSCheckboxItemIndeterminate in an error context and in read only mode")
         }
 
-        if let helper, helper.isEmpty {
-            OL.warning("Helper text given to an OUDSCheckboxItemIndeterminate is defined but empty, is it expected? Prefer use of `nil` value instead")
+        if label.isEmpty {
+            OL.warning("Label given to an OUDSCheckboxItemIndeterminate is defined but empty, prefer OUDSCheckboxIndeterminate(isOn:accessibilityLabel:) instead")
+        }
+
+        if let description, description.isEmpty {
+            OL.warning("Description text given to an OUDSCheckboxItemIndeterminate is defined but empty, is it expected? Prefer use of `nil` value instead")
         }
 
         // swiftlint:disable force_unwrapping
@@ -213,8 +217,8 @@ public struct OUDSCheckboxItemIndeterminate: View {
         self.action = action
         layoutData = .init(
             label: label.localized(),
-            additionalLabel: nil,
-            helper: helper?.localized(),
+            extraLabel: nil,
+            description: description?.localized(),
             icon: icon,
             flipIcon: flipIcon,
             isOutlined: false,
@@ -250,7 +254,7 @@ public struct OUDSCheckboxItemIndeterminate: View {
 
         let checkboxA11yTrait = "core_checkbox_trait_a11y".localized() // Fake trait for Voice Over vocalization
 
-        let result = "\(stateDescription), \(layoutData.label), \(layoutData.helper ?? "") \(errorDescription), \(checkboxA11yTrait)"
+        let result = "\(stateDescription), \(layoutData.label), \(layoutData.description ?? "") \(errorDescription), \(checkboxA11yTrait)"
         return result
     }