+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+
 # OUDS iOS library changelog
 
 All notable changes to this project will be documented in this file.
@@ -5,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased](https://github.com/Orange-OpenSource/ouds-ios/compare/0.22.0..develop)
+
+### Changed
+
+- Update `switch` components to v1.5.0 (Orange-OpenSource/ouds-ios#1138)
+- Update `radio` components to v1.4.0 (Orange-OpenSource/ouds-ios#1139)
+- Update `checkbox` components to v2.4.0 (Orange-OpenSource/ouds-ios#1137)
+- Read only variant for `checkbox` and  `checkbox indeterminate` components (Orange-OpenSource/ouds-ios#1137)
 
 ## [0.22.0](https://github.com/Orange-OpenSource/ouds-ios/compare/0.21.0...0.22.0) - 2025-11-28